Move cmd/dryrun to pkg/dryrun
Because the main package was at `./cmd`, when installed via `go install`
the tool would just be called `cmd`. Now it will be installed as
`dryrun` (although the path is very slightly different).
