+++ cmd/dryrun/main.go
@@ -6,7 +6,7 @@ import (
 	"errors"
 	"os"
 
-	"open-cluster-management.io/config-policy-controller/cmd/dryrun"
+	"open-cluster-management.io/config-policy-controller/pkg/dryrun"
 )
 
 func main() {
