+++ pkg/dryrun/testdata/test_basic_compliant/input_1.yaml
