+++ pkg/dryrun/testdata/test_basic_compliant/policy.yaml
