+++ pkg/dryrun/testdata/test_basic_noncompliant/input_1.yaml
