+++ pkg/dryrun/testdata/test_directory_input/input_directory/input_1-2.yaml
