+++ pkg/dryrun/testdata/test_directory_input/input_directory/input_4.yaml
