+++ pkg/dryrun/testdata/test_directory_input/output.txt
