+++ pkg/dryrun/testdata/test_multiple_resources/input_3.yaml
