+++ pkg/dryrun/testdata/test_multiple_resources/input_4.yaml
