+++ pkg/dryrun/testdata/test_multiple_resources/input_stdin.yaml
