+++ pkg/dryrun/testdata/test_multiple_resources/output.txt
