+++ pkg/dryrun/testdata/test_multiple_resources/policy.yaml
