+++ pkg/dryrun/testdata/test_stdin/input_stdin.yaml
