+++ pkg/dryrun/testdata/test_stdin/output.txt
