+++ pkg/dryrun/testdata/test_stdin/policy.yaml
