+++ pkg/dryrun/testdata/test_wrapped_in_policy/input_1.yaml
